moving into utils/helper.ts
